update and refactor backend routes and deps
