+++ backend/package.json
@@ -18,12 +18,13 @@
     "cookie-parser": "^1.4.6",
     "cookie-session": "^2.1.0",
     "cors": "^2.8.5",
+    "csurf": "^1.10.0",
     "ejs": "^3.1.10",
     "express": "^4.21.0",
     "express-session": "^1.18.0",
     "express-validator": "^7.2.0",
+    "helmet": "^8.0.0",
     "jsonwebtoken": "^9.0.2",
-    "method-override": "^3.0.0",
     "mongoose": "^8.6.2",
     "morgan": "^1.10.0",
     "multer": "^1.4.5-lts.1",
