--- backend/views/clients.ejs
@@ -1,40 +1,0 @@
-<!-- Get all clients for user, and add new clients -->
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Clients</title>
-  </head>
-  <body>
-    <h1>Clients for <%= user.firstName %></h1>
-    <% if (clients.length === 0) { %>
-    <p>No clients found.</p>
-    <% }%>
-    <br />
-    <% clients.forEach((client) => { %>
-    <p><%= client.firstName %></p>
-    <p><%= client.email %></p>
-    <form
-      action="/api/users/<%= user._id %>/clients/<%= client._id %>"
-      method="GET"
-    >
-      <button type="submit">Update Client</button>
-    </form>
-    <form
-      action="/api/users/<%= user._id %>/clients/<%= client._id %>?_method=DELETE"
-      method="POST"
-    >
-      <button type="submit">Delete Client</button>
-    </form>
-    <% }); %>
-    <br />
-    <form action="/api/users/<%= user._id %>/clients" method="POST">
-      <label for="firstName">First Name:</label>
-      <input type="text" name="firstName" id="firstName" />
-      <label for="email">Email:</label>
-      <input type="email" name="email" id="email" />
-      <button type="submit">Add Client</button>
-    </form>
-  </body>
-</html>
