--- backend/views/fileForm.ejs
@@ -1,22 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Document</title>
-  </head>
-  <body>
-    <h1>Test File Form</h1>
-    <form
-      action="/api/projects/<%= projectId %>/uploads"
-      enctype="multipart/form-data"
-      method="POST"
-    >
-      <label for="name">Name:</label>
-      <input type="text" name="name" id="name" />
-      <label for="file">File:</label>
-      <input type="file" name="file" id="file" multiple />
-      <button type="submit">Upload</button>
-    </form>
-  </body>
-</html>
