--- backend/views/index.ejs
@@ -1,16 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Home Page for Sonex Backend</title>
-  </head>
-  <body>
-    <h1>Hi <%= name%>!</h1>
-    <p>This is the home page for Sonex Backend</p>
-
-    <form action="/api/auth/logout?_method=DELETE" method="POST">
-      <button type="submit">Logout</button>
-    </form>
-  </body>
-</html>
