--- backend/views/login.ejs
@@ -1,78 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Sonex Test Login</title>
-    <style>
-      body {
-        font-family: arial;
-        margin: 0;
-      }
-      nav {
-        background: #ff5353;
-        padding: 20px 10px;
-      }
-      nav ul {
-        max-width: 960px;
-        margin: 0 auto;
-        padding: 0;
-      }
-      nav li {
-        list-style-type: none;
-        display: inline-block;
-        margin: 0 10px 0 0;
-      }
-      nav a {
-        color: #fff;
-        font-size: 18px;
-        background: rgba(255, 255, 255, 0.2);
-        text-decoration: none;
-        padding: 10px;
-        display: block;
-      }
-      main,
-      header {
-        max-width: 960px;
-        margin: 30px auto;
-        padding: 0 10px;
-        color: #333;
-      }
-      .google-btn {
-        color: #fff;
-        text-decoration: none;
-        font-size: 18px;
-        padding: 10px;
-        background: #ff5353;
-      }
-    </style>
-  </head>
-  <body>
-    <nav>
-      <ul>
-        <li><a href="/api/auth/logout">Logout</a></li>
-        <li><a href="/api/auth/login">Login</a></li>
-        <li><a href="/">Homepage</a></li>
-      </ul>
-    </nav>
-    <header>
-      <h1>Login using...</h1>
-    </header>
-    <main>
-      <form action="/api/auth/login" method="POST">
-        <div>
-          <label for="username">Username</label>
-          <input type="text" id=username" name="username" required />
-        </div>
-        <div>
-          <label for="password">Password</label>
-          <input type="password" id="password" name="password" required />
-        </div>
-        <button type="submit">Login</button>
-      </form>
-      <br />
-      <a href="/api/auth/login/google" class="google-btn">Google+</a>
-      <a href="/api/auth/register">Sign Up with Email</a>
-    </main>
-  </body>
-</html>
