--- backend/views/projectForm.ejs
@@ -1,67 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Test Update Project Form</title>
-  </head>
-  <body>
-    <h1>Test Update Project Form</h1>
-    <h3>Project: <%= project.title %></h3>
-
-    <form action="/api/projects/<%= project._id %>?_method=PUT" method="POST">
-      <div>
-        <label for="title">Title</label>
-        <input
-          type="text"
-          name="title"
-          id="title"
-          value="<%= project.title %>"
-        />
-      </div>
-      <div>
-        <label for="description">Description</label>
-        <textarea name="description" id="description">
-<%= project.description %></textarea
-        >
-      </div>
-      <div>
-        <label for="status">Status</label>
-        <input
-          type="text"
-          name="status"
-          id="status"
-          value="<%= project.status %>"
-        />
-      </div>
-      <div>
-        <label for="paymentStatus">Payment Status</label>
-        <input
-          type="text"
-          name="paymentStatus"
-          id="paymentStatus"
-          value="<%= project.paymentStatus %>"
-        />
-      </div>
-      <div>
-        <label for="projectAmount">Project Amount</label>
-        <input
-          type="number"
-          name="projectAmount"
-          id="projectAmount"
-          value="<%= project.projectAmount %>"
-        />
-      </div>
-      <div>
-        <label for="amountPaid">Amount Paid</label>
-        <input
-          type="number"
-          name="amountPaid"
-          id="amountPaid"
-          value="<%= project.amountPaid %>"
-        />
-      </div>
-      <button type="submit">Update project</button>
-    </form>
-  </body>
-</html>
