--- backend/views/projects.ejs
@@ -1,54 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Sonex Test Project Form</title>
-  </head>
-  <body>
-    <h1>Create A New Project</h1>
-    <!-- Test Project Form -->
-    <form action="/api/projects" method="POST">
-      <div>
-        <label for="title">Title</label>
-        <input type="text" name="title" id="title" />
-      </div>
-      <div>
-        <label for="description">Description</label>
-        <textarea name="description" id="description"></textarea>
-      </div>
-      <div>
-        <label for="projectAmount">Project Amount</label>
-        <input type="number" name="projectAmount" id="projectAmount" />
-      </div>
-      <button type="submit">Add new project</button>
-    </form>
-    <h3>Projects for <%= user.firstName %></h3>
-    <div id="projects">
-      <% if (projects.length > 0) { %>
-      <ul>
-        <% projects.forEach(function(project) { %>
-        <div>
-          <li><%= project.title %></li>
-          <span><%= project.status %></span>
-          <span>Payment Status: <%= project.paymentStatus %></span>
-          <span>Project Amount: <%= project.projectAmount %></span>
-          <span>Total Paid: <%= project.amountPaid %></span>
-          <form action="/api/projects/<%= project._id %>/update" method="GET">
-            <button type="submit">Update</button>
-          </form>
-          <form
-            action="/api/projects/<%= project._id %>?_method=DELETE"
-            method="POST"
-          >
-            <button type="submit">Delete</button>
-          </form>
-        </div>
-        <% }); %>
-      </ul>
-      <% } else { %>
-      <p>No projects found</p>
-      <% } %>
-    </div>
-  </body>
-</html>
