--- backend/views/register.ejs
@@ -1,78 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Sonex Test Register</title>
-    <style>
-      body {
-        font-family: arial;
-        margin: 0;
-      }
-      nav {
-        background: #ff5353;
-        padding: 20px 10px;
-      }
-      nav ul {
-        max-width: 960px;
-        margin: 0 auto;
-        padding: 0;
-      }
-      nav li {
-        list-style-type: none;
-        display: inline-block;
-        margin: 0 10px 0 0;
-      }
-      nav a {
-        color: #fff;
-        font-size: 18px;
-        background: rgba(255, 255, 255, 0.2);
-        text-decoration: none;
-        padding: 10px;
-        display: block;
-      }
-      main,
-      header {
-        max-width: 960px;
-        margin: 30px auto;
-        padding: 0 10px;
-        color: #333;
-      }
-      .google-btn {
-        color: #fff;
-        text-decoration: none;
-        font-size: 18px;
-        padding: 10px;
-        background: #ff5353;
-      }
-    </style>
-  </head>
-  <body>
-    <h1>Register</h1>
-    <form action="/api/auth/register" method="POST">
-      <div>
-        <label for="firstName">First Name</label>
-        <input type="text" name="firstName" required />
-      </div>
-      <div>
-        <label for="lastName">Last Name</label>
-        <input type="text" name="lastName" />
-      </div>
-      <div>
-        <label for="email">Email</label>
-        <input type="email" name="email" required />
-      </div>
-      <div>
-        <label for="username">Username</label>
-        <input type="text" name="username" required />
-      </div>
-      <div>
-        <label for="password">Password</label>
-        <input type="password" name="password" required />
-      </div>
-      <button type="submit">Register</button>
-    </form>
-
-    <a href="/api/auth/login">Login</a>
-  </body>
-</html>
